apps/listen/pyproject.toml
[project]
name = "listen"
version = "0.1.0"
description = "HTTP server + job manager for remote agent execution"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pyyaml",
]
